requirements.txt
mysql-connector-python
ttkthemes
pillow
tkcalendar
reportlab
